# apps/backend/docs/swagger.yaml
basePath: /
definitions:
  common.ErrorResponse:
    properties:
      error:
        example: Invalid state
        type: string
    type: object
  controllers.AuthCallbackResponse:
    properties:
      message:
        example: Login successful
        type: string
      sub:
        example: auth0|1234567890
        type: string
    type: object
  main.VaultResponse:
    properties:
      message:
        example: Welcome to your secure vault
        type: string
      sub:
        example: auth0|1234567890
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Secure self-hosted document archival API
  title: Enklave API
  version: "1.0"
paths:
  /api/vault:
    get:
      description: Returns secure vault message and subject identity for authenticated
        user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.VaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Access secure vault
      tags:
      - vault
  /auth/callback:
    get:
      description: Handles OIDC authentication callback, verifies token, and creates
        user session
      parameters:
      - description: OAuth2 state parameter
        in: query
        name: state
        required: true
        type: string
      - description: OAuth2 authorization code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/controllers.AuthCallbackResponse'
        "400":
          description: Invalid state
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Authentication or token verification error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: OIDC callback
      tags:
      - auth
  /auth/login:
    get:
      description: Redirects to the OIDC identity provider for authentication
      responses:
        "302":
          description: Redirects to OIDC provider auth URL
          schema:
            type: string
      summary: OIDC login
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer " followed by your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"